fix(omni): address review feedback on storage module
- Deduplicate dirSize: extract to shared export in omni-gc.ts
- Validate extension parameter to prevent path traversal
- Remove misleading resolveObjectPath (findObjectPath is correct API)
- Move OmniStoragePaths interface to types.ts
